Support ceph_keys.sh option 3
Update ceph_keys so it can support all three options below:

1. make control-plane/ceph_keys.yaml with CephExtraKeys
2. make ~/dcn_ceph_keys.yaml with CephExtraKeys and CephExternalMultiConfig
3. make control-plane/ceph_keys_update.yaml with CephExternalMultiConfig

This required adding get_from_yaml.py to extract the necessary
data from the config-download directories of other deployments
for cases 2 and 3 whenver CephExternalMultiConfig needs to be
populated.

In retrospect ceph_keys probably should have just been written
in Python but it's just a helper for me to more easily repeat
this type of deployment.

We'll probably need to supply some sort of tooling to manage
this for TripleO users who are executing this usecase and
this was an exploration of the types of problems which would
need to be solved.