Introduce ansible_cfg.sh
ansible_cfg.sh creates an ansible.cfg in the current users home.
It uses 'openstack tripleo config generate ansible' to do this
but then it updates the ansible.cfg with sed to replace the roles
and library paths to suit my development environment.

Since I tend to checkout tripleo-ansible, tripleo-validations, etc
from git, I have those repositories in my home directory so I want
to use the standard tripleo defaults with the exception of development
copies I am working on. I used to do this by exporting environment
variables in my shell scripts to execution ansible but this is cleaner.