Remove RootStackName
RootStackName should be central on Central stack and dcn1 on DCN1 stack.
RootStackName is set automatically in mistral to the container name.
It shouldn't need to be set manually and it's used in nova-as-config
as the default AZ name keystone to generated clouds.yaml, and
cinder-api to set OS_CLOUD. If an AZ name is explicitly set for the
DCN stack then RootStackName is not used at all.

It was getting set as a workaround to RHBZ 1774731 but shouldn't
be necessary.